Exclude hadoop-client and jars it depends on from packaging.
These need to be filtered out to ensure compatibility with different Hadoop
versions when we need to be able to use HADOOP_USER_CLASSPATH_FIRST and
mapreduce.job.user.classpath.first. Those settings are needed, for example,
since our job needs to use a recent enough version of Jackson, but other jars
put on the classpath by the Hadoop distribution, such as HDP, get in the
way. Without filtering out hadoop-client, we end up with Hadoop 1
incompatibilities as the hadoop dependencies are pulled from our version instead
of the Hadoop distribution's version of the jars.